+++ packages/jds/src/components/Badge/ContentBadge/ContentBadge.tsx
