+++ packages/jds/src/components/Badge/ContentBadge/contentBadge.css.ts
