+++ packages/jds/src/components/Badge/ContentBadge/contentBadge.types.ts
