+++ packages/jds/src/components/Badge/index.ts
@@ -1,11 +1,11 @@
 export { Badge } from "./Badge";
-export * from "./contentBadge/ContentBadge";
-export * from "./dotBadge/DotBadge";
-export * from "./numericBadge/NumericBadge";
+export * from "./ContentBadge/ContentBadge";
+export * from "./DotBadge/DotBadge";
+export * from "./NumericBadge/NumericBadge";
 
 export type { BadgeSize, BasicHierarchy, FeedbackVariant } from "./badge.types";
-export type { ContentBadgeStyle, ThemeVariant } from "./contentBadge/contentBadge.types";
-export type { NumericBadgeStyle } from "./numericBadge/numericBadge.types";
+export type { ContentBadgeStyle, ThemeVariant } from "./ContentBadge/contentBadge.types";
+export type { NumericBadgeStyle } from "./NumericBadge/numericBadge.types";
 
 export {
   BADGE_SIZE_OPTIONS,
@@ -15,20 +15,20 @@ export {
 export {
   CONTENT_BADGE_STYLE_OPTIONS,
   THEME_VARIANT_OPTIONS,
-} from "./contentBadge/contentBadge.types";
-export { NUMERIC_BADGE_STYLE_OPTIONS } from "./numericBadge/numericBadge.types";
+} from "./ContentBadge/contentBadge.types";
+export { NUMERIC_BADGE_STYLE_OPTIONS } from "./NumericBadge/numericBadge.types";
 
 export type {
   ContentBadgeBasicProps,
   ContentBadgeFeedbackProps,
   ContentBadgeProps,
   ContentBadgeThemeProps,
-} from "./contentBadge/contentBadge.types";
+} from "./ContentBadge/contentBadge.types";
 
-export type { DotBadgeFeedbackProps, DotBadgeProps } from "./dotBadge/dotBadge.types";
+export type { DotBadgeFeedbackProps, DotBadgeProps } from "./DotBadge/dotBadge.types";
 
 export type {
   NumericBadgeBasicProps,
   NumericBadgeFeedbackProps,
   NumericBadgeProps,
-} from "./numericBadge/numericBadge.types";
+} from "./NumericBadge/numericBadge.types";
